requirements.txt
pyserial==3.4.0
PyQt5==5.11.3